Database update: removed field "ID_VARIABLE" from "services_variables" table. Unnecessary. Database update: Added field "ENABLED" to tables "actions", "services_actions", "services_operators" and "services_variables" Database update: Added field "ID_SEAT" to tables "operators". Removed field "IS_CALL_CENTER", operators with ID_SEAT=NULL are considered to be call center. Database update: Removed field "SEAT_EXTEN" from table "seats". Extensions will be in a separate table. Database update: Added table "extensions". Each seat can have multiple extensions.
Removed commands "EXTEN" and "IDEN" from client-server protocol. Not needed.
Implemented "SERVICES" and "QUEUES" commands on server side.

Rewritten code to conform to the database changes.
Updated documentation.

+++ AstCTIServer/configurationchecker.h
@@ -73,7 +73,7 @@ signals:
 
 private:
 	Q_DISABLE_COPY(ConfigurationChecker)
-	int                 readLastModified();
+	long                readLastModified();
 	QVariant            readSetting(const QString &name, const QVariant &defaultValue);
 	bool                readConfiguration();
 	QString             timestampToString(const long timestamp);
